move all tests to main_test
